backend/requirements.txt
fastapi
httpx
svgwrite
uvicorn
python-dotenv
